+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * add scrollbar to measurement list box
 * fix id error when column is present in data set with and without a unit specified
 * fix floating point parsing on some locales
+* use system locale for anything except file parsing
 
 ## 1.0.0 (05.03.2025)
 * initial release